Changes bars to be defined as view attributes
The `status_bar` was positioned absolutely regardless of where it was
defined in the view tree. This was a bit messy, so it's been removed in
favor of setting attributes on the `view` element.

The `view` element now supports `top_bar` and `bottom_bar` attributes
which will render a `bar` element at the top or bottom of the view,
respectively.

Bars are single-column containers for text, intended to support status
and menu bars.