Renames text_group to label, only elements as child nodes
`label` is intended to be the main way of displaying arbitrary text. It
has one or more child `text` nodes that actually hold content, and
provides the layout for this content.

The other major change was to how non-element values are stored on an
element. Initially, text elements had a list of binaries as children.
Sparkline elements had a list of numbers as children. This has been
revised so that these sorts of values are stored as attributes instead,
which simplifies recursing on a view tree.