Adds `tree` and `tree_node` elements
This provides support for drawing tree structures in ExTermbox
applications similar to those that the unix tree utility produces.

A `tree` element wraps the structure, but is not a node itself. It
contains zero or more root `tree_node` elements.

Each `tree_node` may have content (given as the `:content` attribute)
and may contain zero or more child `tree_node` elements.

    tree do
      tree_node content: "A" do
        tree_node content: "C" do
        tree_node(content: "E")
        tree_node(content: "F")
        end
      end
      tree_node content: "B" do
        tree_node(content: "D")
      end
    end

There's currently no support for styling the tables or providing
other elements as content, but this may come later.