Standardizes the View DSL
Instead of manually defining the DSL (which led to some inconsistencies
between tags), all supported forms are now automatically generated for
each tag.

Support for passing content directly when defining some elements (e.g.,
labels, text nodes, table rows) was removed. This was previously there
as a shortcut for defining it via an attribute, but it was inconsistent
and may have been misleading. Content should now be defined explicitly
via the relevant attributes:

    text("foo")          # Before
    text(content: "foo") # After

    sparkline([1, 2, 3])         # Before
    sparkline(values: [1, 2, 3]) # After