Improves calculation of column and panel widths
Instead of leaving empty columns on the right edge, we try to increase
the padding between columns if this can be applied evenly.

Fixes an off-by-one error in the box width and height calculations,
which was also leading to underutilization of the available width. A box
with corners (0, 0) and (9, 1) has a width of 10 and a height of 2, but
these were being calculated as 9 and 1, respectively.

Adds tests for various column layouts.