Introduces table_cell tag
Previously the cells of a table row were defined via the `values`
attribute. While this was nice and compact, it only supported plain text
values:

    table_row(values: ["a", "b", "c"])

With the old definition style, it wasn't possible to style one of the
cells differently, or even style content within a cell differently.

This change introduces the `table_cell` tag as the new way of defining
table cells. It doesn't make any changes to styling, and table cells
cannot currently have child elements, but that will both change in the
future.

    table_row do
      table_cell(content: "a")
      table_cell(content: "b")
      table_cell(content: "c")
    end

In the future, this should also be possible:

    table_row do
      table_cell(attribute: @bold) do
        text(color: @red, content: "Red")
        text(color: @blue, content: "Blue")
      end
    end

While the new DSL is a bit more verbose, we get back flexibility in
exchange. If it's too unwieldly in real-world applications, I may
reintroduce `values`.