Adds element specs and attribute validations
Each element type now has a specification defining its tag, which child
tags (if any) it may have, which attributes may be defined on it, and
which of those are required.

It's now an error to use attributes that are not defined in the
element's spec or to exclude those that are required.