Improves the runtime and adds tests for it
Makes some adjustments to shutdown options and the supervisor's restart
policy for the runtime and its deps. Need to be a little careful not to
trigger segfaults in the termbox code by e.g. initializing it again
after it was not cleanly shut down.

The quit events can now be specified via `Runtime.start_link/1`.