Logs raised errors to the terminal after termbox is shut down
We have to rescue errors in the application loop so that we can stop
termbox and ensure they're actually printed out to the console (if
termbox is still active, they disappear into the void and the experience
is frustrating).

Should really improve the development UX, as viewing errors was
difficult before (or required logging to file).

Only supported when using the runtime.