Adds a new `viewport` element (e.g., to implement scrolling)
In the browser, the viewport is the visible portion of the document,
which may extend above, below, left or right of what's currently
visible.

When we scroll in the browser, the viewport itself doesn't move, but
rather the content is being moved into view. It's analagous to trying
to read a large map with your eyes and neck fixed in place. To see
different regions of the map, you need to move the map itself.

The `viewport` element in Ratatouille is based on this concept, but
we're not limited to a single one---they can be arbitrarily placed and
even nested. Elsewhere, it might be called a "scrollable" container,
but I thought that might confuse users into thinking it's somehow
dynamic. Ratatouille's views are always rendered statically, and the
only way to change what's rendered is to render a new view.

With that in mind, the viewport is a static view of its child content
with a particular x and y offset. It's useful when the content is
larger than the viewport itself. By default, the top-left corner of
the content is rendered to the top-left corner of the viewport.
Increasing the x offset by 1 hides the left-most column from view.
Likewise, increasing the y offset by 1 hides the top-most row from
view. In this way, the offsets allow us to move the content behind the
viewport.

While scrolling isn't implemented here, this makes it possible to
implement. We can generate new views and iteratively increase the
viewport's offset until an edge is reached.

    def init(model), do: %{cursor: 0}

    def update(model, message) do
      case message do
        {:event, %{key: @arrow_down}} ->
	  %{model | cursor: model.cursor + 1}
      end
    end

    def render(model) do
      viewport(offset_y: model.cursor) do
        # ... lots of content ...
      end
    end

Although the viewport concept itself seems fine, I think there are a
few pieces missing to properly implement scrolling:

* It's not possible to dynamically retrieve the dimensions of child
  content from a Ratatouille application, so there's no sure-fire way
  to determine where the cursor should end. We need a sort of API to
  query the dimensions of rendered content.
* Particularly with tables, child content may be spaced or rendered
  differently differently based on the available rendering space. The
  viewport works by extending the rendering region and then slicing
  off everything before the offset. So this can lead to the content
  stretching as the offset (and rendering region) grows. There are
  different ways to fix this, but each have different tradeoffs.

The viewport element's own height and width should be configurable,
but currently it just occupies all the available space remaining in
the parent container. I'll add the height and width configuration in
the future.

Closes #9